+++ CHANGELOG.md
@@ -14,7 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New logger to send logs to Grafana Cloud service.
 - New listener for bbPress.
 - New listener for BuddyPress.
-- New listeners for Stripe gateways (standard, Amelia and Forminator).
+- New listener for Action Scheduler library.
+- New listeners for Stripe gateways (standard, Amelia and Forminator) libraries.
 - New 'environment' context metadata for all external loggers.
 - Compatibility with WordPress 5.7.
 - New setting to override live console and local access privileges when in development or staging environments (thanks to [sebastienserre](https://github.com/sebastienserre) for the suggestion).